Move estimateTxSize etc to separate module
