+++ lib/wallet/cardano-wallet.cabal
@@ -409,6 +409,7 @@ library
     Cardano.Wallet.Write.ProtocolParameters
     Cardano.Wallet.Write.Tx
     Cardano.Wallet.Write.Tx.Balance
+    Cardano.Wallet.Write.Tx.Balance.CoinSelection
     Cardano.Wallet.Write.Tx.Gen
     Cardano.Wallet.Write.Tx.Redeemers
     Cardano.Wallet.Write.Tx.TimeTranslation
